+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/oom/keys/Metric.java
@@ -2,7 +2,7 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.lang.java.oom;
+package net.sourceforge.pmd.lang.java.oom.keys;
 
 /**
  * Umbrella marker interface.
