+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/oom/keys/MetricKey.java
@@ -1,0 +1,19 @@
+/**
+ * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
+ */
+
+package net.sourceforge.pmd.lang.java.oom.keys;
+
+/**
+ * Key identifying a metric.
+ */
+public interface MetricKey {
+
+    /**
+     * Returns the name of the metric.
+     *
+     * @return The name of the metric.
+     */
+    String name();
+
+}
